Make sure to update GENERIC_PATCH_DIR when generating a checkpoint
GENERIC_PATCH_DIR is normally calculated by quilt.mk in order to figure out
which kernel patches to use. For a checkpoint it uses a static patches
directory. This commit overrides the logic which calculates the value of
GENERIC_PATCH_DIR by adding it to the config file at checkpoint creation.

Signed-off-by: Behan Webster <behanw@converseincode.com>