Add -Wgnu to kernel Makefile
There are a lot of warnings about zero-length VLAs not being at the end of
structures. They are in structures which are used in other structures. However,
since it is merely a warning about the use of a gnu extension which otherwise
still compiles and works, the warning is merely annoying and not ultimately
useful.

Signed-off-by: Behan Webster <behanw@converseincode.com>