Add kernel build shell feature
Configure and launch an interactive shell to build LLVMLinux kernels.

This allows one to setup LLVMLinux infrastructure in her command prompt and
build Linux kernels as it is done in a normal kernel setup. Hence one will be
able to generate .s and .o files and use the KConfig facilities such as
'make modules_install'. This can be very handy for debugging purposes.

The commands 'make kernel-shell-for-build' or 'kernel-gcc-shell-for-build'
will print in one line all variables needed for building the kernels.

Usage
  # Select one target
  cd targets/nexus7
  # Configure the kernel
  make kernel-configure
  # Create a new shell
  eval `make -s kernel-shell-for-build` $SHELL
  # Build an specific target
  $BUILD_COMMAND fs/cfq_iosched.o
  # Build all
  $BUILD_COMMAND
  # Leave
  exit

Notice that 'kernel-gcc-configure' is needed by 'kernel-gcc-shell-for-build'.

Signed-off-by: Vinícius Tinti <viniciustinti@gmail.com>