Simplied vlais-netfilter patch and updated unit test
The new patch leaves the entries structure member as a
flexible array. Now only the term member needs to have
its offset calculated.

The padding between tbl->entries[nhooks] and tbl->term
must be calculated as well as the parring after term to
maintain the same memory layout as the VLAIS version.

Signed-off-by: Mark Charlebois <charlebm@gmail.com>