Needed to add -target to CCOPTS for aarch86 for libgcc
AARCHC64 kernel depends on libgcc and needs to be able
to find it via $(CC) -print-libgcc-file-name. This is
potentially a problem if compiler-rt was to be used in
place of libgcc.

Addint the -target option to clang allows the
-print-libgcc-file-name to resolve the path to libgcc.

inline-asm-aarch64.patch was also updated with more
fixes for improper uses of 32 bit vars in asm calls
for 64 bit registers.

Since vexpress64 switched from the linaro tree to the
mainline kernel, the master directory is no longer needed
to differentiate the trees.

Signed-off-by: Mark Charlebois <charlebm@gmail.com>