Added support for cubieboard
Cubieboard is supported in the upstream kernel and will build.
There may still be some additional patches required for it to run.

QEMU v2.0 supports the -M cubieboard option now so QEMU is now
build from the master branch.

The kernel builds successfully as does QEMU but the resulting kernel
does not boot in QEMU.

Signed-off-by: Mark Charlebois <charlebm@gmail.com>