Use fakeroot to create file nodes in initramfs
The kernel config file now points to the generated
initramfs.cpio.gz that contains the /dev/console and
/dev/tty0. Kernels were unable to open the console if
these nodes did not exist before init was run.

Includes patch to initramfs.mk from Daniel.Sanders@imgtec.com:

The ${BUSYBOXURL} values have an architecture suffix but the build
framework expects the resulting filename to be 'busybox'.

This patch fixes that so that the copy in ${INITTMPDIR} is expected to have
a suffix, and it is removed when copying to ${INITBUILDDIR}.

Signed-off-by: Mark Charlebois <charlebm@gmail.com>