Added initramfs for aarch64 using Linaro rootfs
There are no prebuld static versions of busybox for aarch64
that support --install.

It is not possible to build recent versions of busybox standalone
and have it be full featured so a shared libc is required. Using
the minimal Linaro rootfs for aarch64 as a initramfs allows
testing and has minimal setup overhead.

Signed-off-by: Mark Charlebois <charlebm@gmail.com>