patch: replace hack for trusted foundations
As documented in GCC naked functions should only use Basic asm syntax. The
Extended asm or mixture of Basic asm and "C" code is not guarantee. Currently
this works because it was hard coded to follow and check GCC optimization
calling convention.

This patch removes the hard code usage and checks and instead makes the naked
function to be inline which will place the code in the same way as the naked
function.

https://gcc.gnu.org/onlinedocs/gcc/Function-Attributes.html

Signed-off-by: Behan Webster <behanw@converseincode.com>
Signed-off-by: Vinícius Tinti <viniciustinti@gmail.com>