patch: disable register allocation patch
The patch c097877319ab61dd045b6497953b4e3df8f2bb44 in Linux move psci calls to
an assembly file and by doing so remove the '__asmeq' calls. But I have not seen
any guarantee that the register placement will hold.

This can be a potential issue in the future.

Signed-off-by: Vinicius Tinti <viniciustinti@gmail.com>