[mips][malta] Added support for core boards and I6400.
Malta is a development kit with a swappable 'core board'. So far we've only
supported a generic MIPS32r2. This patch adds support for core boards
and adds the I6400 core board which will exercise the MIPS32r6 (and later
the MIPS64r6) portions of the mips arch.

The I6400 support doesn't boot yet but the build system is done.

Signed-off-by: Daniel Sanders <daniel.sanders@imgtec.com>